tests_vechain/requirements_vechain.txt
pytest==6.2.5
python-decouple==3.6
thor-requests==1.2.2

